+++ commons/commons-model-extended-test/build.gradle.kts
@@ -4,7 +4,4 @@ import Projects.CommonsModelExtended
 dependencies {
     api(project(CommonsModelExtended))
     implementation(Arrow)
-
-    testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.0")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine")
 }
