+++ submission/persistence-mongo/src/test/kotlin/ac/uk/ebi/biostd/persistence/doc/service/SubmissionMongoPersistenceServiceTest.kt
@@ -1,16 +1,16 @@
 package ac.uk.ebi.biostd.persistence.doc.service
 
-import DefaultSubmission.Companion.ACC_NO
-import DefaultSubmission.Companion.defaultSubmission
+import SubmissionFactory.ACC_NO
+import SubmissionFactory.defaultSubmission
 import ac.uk.ebi.biostd.persistence.common.request.SaveSubmissionRequest
 import ac.uk.ebi.biostd.persistence.doc.db.data.SubmissionDocDataRepository
 import ac.uk.ebi.biostd.persistence.doc.db.data.SubmissionRequestDocDataRepository
 import ac.uk.ebi.biostd.persistence.doc.model.SubmissionRequest
-import ac.uk.ebi.biostd.persistence.doc.model.SubmissionRequestStatus
+import ac.uk.ebi.biostd.persistence.doc.model.SubmissionRequestStatus.REQUESTED
 import ac.uk.ebi.biostd.persistence.filesystem.request.FilePersistenceRequest
 import ac.uk.ebi.biostd.persistence.filesystem.service.FileSystemService
 import com.mongodb.BasicDBObject.parse
-import ebi.ac.uk.extended.model.ExtProcessingStatus.REQUESTED
+import ebi.ac.uk.extended.model.ExtProcessingStatus
 import ebi.ac.uk.extended.model.FileMode.COPY
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
@@ -41,19 +41,21 @@ class SubmissionMongoPersistenceServiceTest(
 
     @Nested
     inner class SaveRequest {
+        private val serializedSub = "{}"
+
         @Test
         fun `save request with current version active`() {
             val subRequestSlot = slot<SubmissionRequest>()
             val current = defaultSubmission(version = 1)
-            val newVersion = defaultSubmission(version = 2, status = REQUESTED)
+            val newVersion = defaultSubmission(version = 2, status = ExtProcessingStatus.REQUESTED)
             every { subDataRepository.getCurrentVersion(ACC_NO) } returns 1
-            every { serializationService.serialize(newVersion, Properties(true)) } returns "{}"
+            every { serializationService.serialize(newVersion, Properties(true)) } returns serializedSub
             every { submissionRequestDocDataRepository.saveRequest(capture(subRequestSlot)) } answers { nothing }
 
             val result = testInstance.saveSubmissionRequest(current)
 
             assertThat(result).isEqualTo(newVersion)
-            val submissionRequest = SubmissionRequest(ACC_NO, 2, SubmissionRequestStatus.REQUESTED, parse("{}"))
+            val submissionRequest = SubmissionRequest(ACC_NO, 2, REQUESTED, parse(serializedSub))
             assertThat(subRequestSlot.captured).isEqualToIgnoringGivenFields(submissionRequest, "id")
         }
 
@@ -61,15 +63,15 @@ class SubmissionMongoPersistenceServiceTest(
         fun `save request with current version deleted`() {
             val subRequestSlot = slot<SubmissionRequest>()
             val current = defaultSubmission(version = -1)
-            val newVersion = defaultSubmission(version = 2, status = REQUESTED)
+            val newVersion = defaultSubmission(version = 2, status = ExtProcessingStatus.REQUESTED)
             every { subDataRepository.getCurrentVersion(ACC_NO) } returns -1
-            every { serializationService.serialize(newVersion, Properties(true)) } returns "{}"
+            every { serializationService.serialize(newVersion, Properties(true)) } returns serializedSub
             every { submissionRequestDocDataRepository.saveRequest(capture(subRequestSlot)) } answers { nothing }
 
             val result = testInstance.saveSubmissionRequest(current)
 
             assertThat(result).isEqualTo(newVersion)
-            val submissionRequest = SubmissionRequest(ACC_NO, 2, SubmissionRequestStatus.REQUESTED, parse("{}"))
+            val submissionRequest = SubmissionRequest(ACC_NO, 2, REQUESTED, parse(serializedSub))
             assertThat(subRequestSlot.captured).isEqualToIgnoringGivenFields(submissionRequest, "id")
         }
 
@@ -77,15 +79,15 @@ class SubmissionMongoPersistenceServiceTest(
         fun `save request without current version`() {
             val subRequestSlot = slot<SubmissionRequest>()
             val current = defaultSubmission(version = 1)
-            val newVersion = defaultSubmission(version = 1, status = REQUESTED)
+            val newVersion = defaultSubmission(version = 1, status = ExtProcessingStatus.REQUESTED)
             every { subDataRepository.getCurrentVersion(ACC_NO) } returns null
-            every { serializationService.serialize(newVersion, Properties(true)) } returns "{}"
+            every { serializationService.serialize(newVersion, Properties(true)) } returns serializedSub
             every { submissionRequestDocDataRepository.saveRequest(capture(subRequestSlot)) } answers { nothing }
 
             val result = testInstance.saveSubmissionRequest(current)
 
             assertThat(result).isEqualTo(newVersion)
-            val submissionRequest = SubmissionRequest(ACC_NO, 1, SubmissionRequestStatus.REQUESTED, parse("{}"))
+            val submissionRequest = SubmissionRequest(ACC_NO, 1, REQUESTED, parse(serializedSub))
             assertThat(subRequestSlot.captured).isEqualToIgnoringGivenFields(submissionRequest, "id")
         }
     }
